+++ subtractor/src/lib.rs
@@ -1,11 +1,12 @@
 mod bindings;
 
+use crate::bindings::calc::basic::types::Numeric;
 use crate::bindings::exports::calc::basic::subtract::Guest;
 
 struct Component;
 
 impl Guest for Component {
-    fn subtract(a: f64, b: f64) -> f64 {
+    fn subtract(a: Numeric, b: Numeric) -> Numeric {
         a - b
     }
 }
